refactor : 인수 최소한으로 전달하도록 코드 수정, 들여쓰기 스트림 통해서 줄이기
